pacientes GET - se crea funcion MATCHING y se pasa la funcionalidad a controller
